(PS3) Cut down on warnings

+++ libretro-common/include/streams/interface_stream.h
@@ -76,7 +76,7 @@ int intfstream_seek(intfstream_internal_t *intf,
 
 void intfstream_rewind(intfstream_internal_t *intf);
 
-size_t intfstream_tell(intfstream_internal_t *intf);
+int intfstream_tell(intfstream_internal_t *intf);
 
 void intfstream_putc(intfstream_internal_t *intf, int c);
